:broom: refactor _get_name
